Allow shifted constrain ranges in dynamo._export
Summary: Previously, when we had two dynamic shape symbols `s0` and `s1` bound by the relationship `s1 == s0 + 1`, even when the range constraints were set in accordance with the relationship, `torch._dynamo.export` raised the error saying that the constraint is violated. Here we add a range check between the expression and the constraint and, if the ranges match, don't declare the constraint violated.

We also add a flag to disable the dim constraint solver in `torch.dynamo._export` (not set by default for BC). This is because, even for simple constraints like `s1 == s0 + 1`, the solver claims that the constraints are too complex and the dimension `s0` must be specialized.

Both changes are needed to unblock PT2 compilation of an internal model.

Test Plan:

```
$ python test/inductor/test_aot_inductor.py -k test_shifted_constraint_ranges
...
----------------------------------------------------------------------
Ran 4 tests in 53.247s

OK (skipped=1)
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 8447e69015231dcae7807244a5cda9d9b09787a6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114024